refactor: normalization directory structure
