+++ src/web/normalization/checkboxHtmlNormalizer.ts
