+++ src/web/useOnChangeHtml.ts
@@ -2,7 +2,7 @@ import { type Editor } from '@tiptap/react';
 import type { OnChangeHtmlEvent } from '../types';
 import type { NativeSyntheticEvent } from 'react-native';
 import { useOnEditorChange } from './useOnEditorChange';
-import { normalizeHtmlFromTiptap } from './tiptapHtmlNormalizer';
+import { normalizeHtmlFromTiptap } from './normalization/tiptapHtmlNormalizer';
 
 export const useOnChangeHtml = (
   editor: Editor,
